Moved function to Util.
