+++ assets/js/util.ts
@@ -3,6 +3,7 @@ import {
   CustomMessage,
 } from "./components/OutfrontTakeoverTool/OutfrontTakeoverTool";
 import CANNED_MESSAGES from "./constants/messages";
+import STATION_ORDER_BY_LINE from "./constants/stationOrder";
 import STATIONS_BY_LINE from "./constants/stations";
 import { Alert } from "./models/alert";
 import { Place } from "./models/place";
@@ -195,4 +196,24 @@ export const sortScreens = (screenList: Screen[]) => {
   return screenList.sort((a, b) =>
     screenTypeOrder.indexOf(a.type) >= screenTypeOrder.indexOf(b.type) ? 1 : -1
   );
+};
+
+export const sortByStationOrder = (
+  places: Place[],
+  filteredLine: string,
+  reverse?: boolean
+) => {
+  const stationOrder = STATION_ORDER_BY_LINE[filteredLine.toLowerCase()];
+
+  places.sort((placeA, placeB) => {
+    const indexA = stationOrder.findIndex((station) => {
+      return station.name.toLowerCase() === placeA.name.toLowerCase();
+    });
+    const indexB = stationOrder.findIndex((station) => {
+      return station.name.toLowerCase() === placeB.name.toLowerCase();
+    });
+    return indexA - indexB;
+  });
+
+  return reverse ? places.reverse() : places;
 };
